+++ NEWS.md
@@ -4,14 +4,16 @@
 * The following tidyverse functions are now automatically converted to their tidytable equivalents when
 used inside of tidytable's mutate/arrange/filter/summarize/slice.
   + Functions that previously failed but now work:
-    - `n()` --> `n.()`/`.N`
-    - `row_number()` --> `row_number.()`/`1:.N`
+    - `n()` --> `n.()`
+    - `row_number()` --> `row_number.()`
     - `c_across()` --> `c_across.()`
   + Functions that previously worked with tidytable but are now converted to their faster tidytable equivalents:
     - `if_else()`/`ifelse()` --> `ifelse.()`
     - `case_when()` --> `case_when.()`
     - `desc()` --> `desc.()`
-  
+    - `replace_na()` --> `replace_na.()`
+* `replace_na.()`: Now can replace `NULL` values in list-columns
+
 #### Functions with notable speed improvements
 * `mutate_rowwise.()` when `c_across.()` is used